finished the photo part of the user.
When user changes their profile picture, it gets uploaded to servers. Then when user logs in or restarts app, the picture is loaded up into the proper spot.

At some point I will figure out how to store the pic on the device so that the picture can load up even if the user is not connected to the internet.